requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
flask
requests
ipykernel
jupyter
pytest
langchain 
sentence-transformers 
faiss-cpu